requirements.txt
aiogram>=3.4.1,<4.0.0

asyncpg>=0.29.0,<1.0.0

python-dotenv>=1.0.0,<2.0.0

openpyxl>=3.1.2,<4.0.0